+++ src/types/dict.d.ts
@@ -1,18 +1,19 @@
 import { IDictDetailData, IDictDetailDtoData } from "./dictDetail";
+import { IPageable } from './base'
 
 export interface IDictSmallDtoData {
     id: number
 }
 
 export interface IDictData {
-    id: string
-    dictDetails: IDictDetailData[]
-    name: string
-    description: string
+    id?: number
+    dictDetails?: IDictDetailData[]
+    name?: string
+    description?: string
 }
 
-export interface IDictQueryData {
-    blurry: string
+export interface IDictQueryData extends IPageable {
+    blurry?: string
 }
 
 export interface IDictDtoData {