+++ src/views/system/dict/index.vue
@@ -1,133 +1,258 @@
 <template>
   <div class="app-container">
     <!--表单组件-->
-    <el-dialog append-to-body :close-on-click-modal="false" :before-close="crud.cancelCU" :visible="crud.status.cu > 0" :title="crud.status.title" width="500px">
-      <el-form ref="form" :model="form" :rules="rules" size="small" label-width="80px">
-        <el-form-item label="字典名称" prop="name">
-          <el-input v-model="form.name" style="width: 370px;" />
-        </el-form-item>
-        <el-form-item label="描述">
-          <el-input v-model="form.description" style="width: 370px;" />
-        </el-form-item>
-      </el-form>
-      <div slot="footer" class="dialog-footer">
-        <el-button type="text" @click="crud.cancelCU">取消</el-button>
-        <el-button :loading="crud.status.cu === 2" type="primary" @click="crud.submitCU">确认</el-button>
-      </div>
-    </el-dialog>
-    <!-- 字典列表 -->
+    <eForm ref="form" :is-add="isAdd" />
     <el-row :gutter="10">
-      <el-col :xs="24" :sm="24" :md="10" :lg="11" :xl="11" style="margin-bottom: 10px">
+      <el-col
+        :xs="24"
+        :sm="24"
+        :md="10"
+        :lg="10"
+        :xl="10"
+        style="margin-bottom: 10px"
+      >
         <el-card class="box-card">
+          <div slot="header" class="clearfix">
+            <span>字典列表</span>
+            <el-button
+              v-permission="['ADMIN', 'DICT_ALL', 'DICT_CREATE']"
+              class="filter-item"
+              size="mini"
+              style="float: right; padding: 4px 10px"
+              type="primary"
+              icon="el-icon-plus"
+              @click="
+                $refs.form.dialog = true;
+                isAdd = true;
+              "
+              >新增</el-button
+            >
+          </div>
           <!--工具栏-->
           <div class="head-container">
-            <div v-if="crud.props.searchToggle">
-              <!-- 搜索 -->
-              <el-input v-model="query.blurry" clearable size="small" placeholder="输入名称或者描述搜索" style="width: 200px;" class="filter-item" @keyup.enter.native="crud.toQuery" />
-              <rrOperation />
-            </div>
-            <crudOperation :permission="permission" />
+            <!-- 搜索 -->
+            <el-input
+              v-model="query.value"
+              clearable
+              placeholder="输入名称或者描述搜索"
+              style="width: 200px"
+              class="filter-item"
+              @keyup.enter.native="toQuery"
+            />
+            <el-button
+              class="filter-item"
+              size="mini"
+              type="success"
+              icon="el-icon-search"
+              @click="toQuery"
+              >搜索</el-button
+            >
           </div>
           <!--表格渲染-->
-          <el-table ref="table" v-loading="crud.loading" :data="crud.data" highlight-current-row style="width: 100%;" @selection-change="crud.selectionChangeHandler" @current-change="handleCurrentChange">
-            <el-table-column type="selection" width="55" />
-            <el-table-column :show-overflow-tooltip="true" prop="name" label="名称" />
-            <el-table-column :show-overflow-tooltip="true" prop="description" label="描述" />
-            <el-table-column v-permission="['admin','dict:edit','dict:del']" label="操作" width="130px" align="center" fixed="right">
+          <el-table
+            v-loading="loading"
+            :data="data"
+            size="small"
+            highlight-current-row
+            style="width: 100%"
+            @current-change="handleCurrentChange"
+          >
+            <el-table-column
+              :show-overflow-tooltip="true"
+              prop="name"
+              label="名称"
+            />
+            <el-table-column
+              :show-overflow-tooltip="true"
+              prop="description"
+              label="描述"
+            />
+            <el-table-column
+              v-if="
+                checkPermission([
+                  'ADMIN',
+                  'DICT_ALL',
+                  'DICT_EDIT',
+                  'DICT_DELETE',
+                ])
+              "
+              label="操作"
+              width="130px"
+              align="center"
+              fixed="right"
+            >
               <template slot-scope="scope">
-                <udOperation
-                  :data="scope.row"
-                  :permission="permission"
+                <el-button
+                  v-permission="['ADMIN', 'DICT_ALL', 'DICT_EDIT']"
+                  size="mini"
+                  type="primary"
+                  icon="el-icon-edit"
+                  @click="edit(scope.row)"
                 />
+                <el-popover
+                  v-permission="['ADMIN', 'DICT_ALL', 'DICT_DELETE']"
+                  :ref="scope.row.id"
+                  placement="top"
+                  width="180"
+                >
+                  <p>此操作将删除字典与对应的字典详情，确定要删除吗？</p>
+                  <div style="text-align: right; margin: 0">
+                    <el-button
+                      size="mini"
+                      type="text"
+                      @click="$refs[scope.row.id].doClose()"
+                      >取消</el-button
+                    >
+                    <el-button
+                      :loading="delLoading"
+                      type="primary"
+                      size="mini"
+                      @click="subDelete(scope.row.id)"
+                      >确定</el-button
+                    >
+                  </div>
+                  <el-button
+                    slot="reference"
+                    type="danger"
+                    icon="el-icon-delete"
+                    size="mini"
+                  />
+                </el-popover>
               </template>
             </el-table-column>
           </el-table>
           <!--分页组件-->
-          <pagination />
+          <el-pagination
+            :total="total"
+            :current-page="page + 1"
+            style="margin-top: 8px"
+            layout="total, prev, pager, next, sizes"
+            @size-change="sizeChange"
+            @current-change="pageChange"
+          />
         </el-card>
       </el-col>
-      <!-- 字典详情列表 -->
-      <el-col :xs="24" :sm="24" :md="14" :lg="13" :xl="13">
+      <el-col :xs="24" :sm="24" :md="14" :lg="14" :xl="14">
         <el-card class="box-card">
           <div slot="header" class="clearfix">
             <span>字典详情</span>
             <el-button
-              v-if="checkPermission(['admin','dict:add']) && this.$refs.dictDetail && this.$refs.dictDetail.query.dictName"
+              v-if="
+                checkPermission(['ADMIN', 'DICT_ALL', 'DICT_CREATE']) &&
+                this.$refs.dictDetail &&
+                this.$refs.dictDetail.dictName
+              "
               class="filter-item"
               size="mini"
-              style="float: right;padding: 4px 10px"
+              style="float: right; padding: 4px 10px"
               type="primary"
               icon="el-icon-plus"
-              @click="$refs.dictDetail && $refs.dictDetail.crud.toAdd()"
-            >新增</el-button>
+              @click="
+                $refs.dictDetail.$refs.form.dialog = true;
+                $refs.dictDetail.isAdd = true;
+              "
+              >新增</el-button
+            >
           </div>
-          <dictDetail ref="dictDetail" :permission="permission" />
+          <dictDetail ref="dictDetail" />
         </el-card>
       </el-col>
     </el-row>
   </div>
 </template>
 
-<script>
-import checkPermission from '@/utils/permission'
-import dictDetail from './dictDetail'
-import crudDict from '@/api/system/dict'
+<script lang="ts">
+import { Vue, Component, Prop } from "vue-property-decorator";
+import { checkPermission } from "@/utils/permission";
+import initData from "@/mixins/initData";
+import { del } from "@/api/system/dict";
+import dictDetail from "../dictDetail/Index.vue";
+import eForm from "./Form.vue";
+import { mixins } from "vue-class-component";
+import { IDictQueryData, IDictDtoData } from "@/types/dict";
 
-import CRUD, { presenter, header, form } from '@crud/crud'
-import crudOperation from '@crud/CRUD.operation'
-import pagination from '@crud/Pagination'
-import rrOperation from '@crud/RR.operation'
-import udOperation from '@crud/UD.operation'
+interface IDictSearch {
+  value?: string;
+}
 
-const defaultForm = { id: null, name: null, description: null, dictDetails: [] }
-
-export default {
-  name: 'Dict',
-  components: { crudOperation, pagination, rrOperation, udOperation, dictDetail },
-  cruds() {
-    return [
-      CRUD({ title: '字典', url: 'api/dict', crudMethod: { ...crudDict }})
-    ]
+@Component({
+  name: "Dict",
+  components: {
+    dictDetail,
+    eForm,
   },
-  mixins: [presenter(), header(), form(defaultForm)],
-  data() {
-    return {
-      queryTypeOptions: [
-        { key: 'name', display_name: '字典名称' },
-        { key: 'description', display_name: '描述' }
-      ],
-      rules: {
-        name: [
-          { required: true, message: '请输入名称', trigger: 'blur' }
-        ]
-      },
-      permission: {
-        add: ['admin', 'dict:add'],
-        edit: ['admin', 'dict:edit'],
-        del: ['admin', 'dict:del']
-      }
+})
+export default class extends mixins<
+  initData<IDictQueryData, IDictDtoData, IDictSearch>
+>(initData) {
+  private delLoading = false;
+  private checkPermission = checkPermission;
+  private queryTypeOptions = [
+    { key: "name", display_name: "字典名称" },
+    { key: "remark", display_name: "描述" },
+  ];
+
+  created() {
+    this.$nextTick(() => {
+      this.init();
+    });
+  }
+
+  beforeInit() {
+    this.url = "api/dict";
+    const sort = "id,desc";
+    this.params = {
+      page: this.page,
+      size: this.size,
+      sort: sort,
+      blurry: this.query.value,
+    };
+
+    if (this.$refs.dictDetail) {
+      (this.$refs.dictDetail as dictDetail).data = [];
+      (this.$refs.dictDetail as dictDetail).dictName = "";
     }
-  },
-  methods: {
-    checkPermission,
-    // 获取数据前设置好接口地址
-    [CRUD.HOOK.beforeRefresh]() {
-      if (this.$refs.dictDetail) {
-        this.$refs.dictDetail.query.dictName = ''
-      }
-      return true
-    },
-    // 选中字典后，设置字典详情数据
-    handleCurrentChange(val) {
-      if (val) {
-        this.$refs.dictDetail.query.dictName = val.name
-        this.$refs.dictDetail.dictId = val.id
-        this.$refs.dictDetail.crud.toQuery()
-      }
+    return true;
+  }
+
+  private async subDelete(id: number) {
+    this.delLoading = true;
+    try {
+      await del([id]);
+      this.delLoading = false;
+      (this.$refs[id] as any).doClose();
+      this.dleChangePage();
+      this.init();
+      this.$notify({
+        title: "删除成功",
+        type: "success",
+        duration: 2500,
+        message: "",
+      });
+    } catch (err) {
+      this.delLoading = false;
+      (this.$refs[id] as any).doClose();
+      console.log(err.response.data.message);
     }
+  }
+
+  private handleCurrentChange(val?: IDictDtoData) {
+    if (val) {
+      (this.$refs.dictDetail as dictDetail).dictName = val.name;
+      (this.$refs.dictDetail as dictDetail).dictId = val.id;
+      (this.$refs.dictDetail as dictDetail).init();
+    }
+  }
+
+  private edit(data: IDictDtoData) {
+    this.isAdd = false;
+    const dictForm = this.$refs.form as eForm;
+    dictForm.form = {
+      id: data.id,
+      name: data.name,
+      description: data.description,
+    };
+    dictForm.dialog = true;
   }
 }
 </script>
-
-<style scoped>
-</style>